# [Core] Allow more PENDING jobs to be scheduled concurrently (1.4x faster)