@Michaelvll 2024-11-11
We should think of the tradeoff of losing the strict FIFO vs the time spend for scheduling, especially that #4318 has already significantly speed up the job scheduling.